Use the systems' stacksize.
A new function which can be called if a library function is not implemented
(instead of NULL).

Use AROS_ALIGN() to align pointers

Use new macros to access library jumptables.

Added support for stackchecking code in dispatch.s. The boot tasks' valid stack
range is the whole memory (from NULL to (APTR)~0UL).

The code to set the initial stackpointer is in exec/addtask.c

Formatting

Make sure multitasking has a liftoff

Return a distinct value, so the developer knows where the code has exited.


git-svn-id: https://svn.aros.org:8080/svn/aros/trunk/AROS@587 fb15a70f-31f2-0310-bbcc-cdcc74a49acc