In OM_SET/OM_UPDATE method rerender if RetVal of DoSuperMethod was != 0.
In layout method set DTSpecialInfo->si_VertUnit/si_VisVert/
si_TopVert and si_HorizUnit/si_VisHoriz/si_TotHoriz manually.

The reason why this might not be needed in AmigaOS is the way
the MultiView application is coded. There the interconnection
objects are organized in such a way that when doing NotifyAttrs()
in your datatype class this will also come back as an OM_UPDATE
to yourself (--> see ClipView src on Amiga Developer CD). But
this is not the case in AROS MultiView at the moment.


git-svn-id: https://svn.aros.org:8080/svn/aros/trunk/AROS@12797 fb15a70f-31f2-0310-bbcc-cdcc74a49acc