Fixed so that the addon commands work with the Wanderer toolbar (open shell, parent drawer and so on,new drawer..). It has overridden default behaviour so please look at the TODO: comments.
The reason why the wanderer windows with a toolbar didn't get synced 
with the wanderer templates, was that the current dir path was stored in
the windows' MenuStrip objects - which currently can't be changed in Zune.
OM_GET and OM_SET methods on MUIA_Window_Menustrip isn't implemented..

So now I get the current dir path from: 
Object *win = ( Object *)XGET( app, MUIA_Wanderer_ActiveWindow );
STRPTR dr = XGET( win, MUIA_IconWindow_Drawer );

I also set the MenuStrip generation in a separate func so that later,
when we DO have control over it, we can overwrite the old MenuStrip
easily with that func.



git-svn-id: https://svn.aros.org:8080/svn/aros/trunk/AROS@25200 fb15a70f-31f2-0310-bbcc-cdcc74a49acc